+++ tesseraql-identity/src/main/resources/io/tesseraql/identity/pack/default/sql/delete-user-attribute.sql
@@ -1,0 +1,3 @@
+delete from tql_user_attributes
+where user_id = /* userId */ 'u1' and name = /* name */ 'department'
+;